+++ build_tools/services.rb
@@ -8,7 +8,7 @@ module BuildTools
     MANIFEST_PATH = File.expand_path('../../services.json', __FILE__)
 
     # Minimum `aws-sdk-core` version for eventstream support
-    EVENTSTREAM_CORE_VERSION = "3.21.0"
+    EVENTSTREAM_CORE_VERSION = "3.21.1"
     EVENTSTREAM_PLUGIN = "Aws::Plugins::EventStreamConfiguration"
 
     # @option options [String] :manifest_path (MANIFEST_PATH)