+++ gems/aws-sdk-core/lib/seahorse/client/http/response.rb
@@ -10,7 +10,6 @@ module Seahorse
           @status_code = options[:status_code] || 0
           @headers = options[:headers] || Headers.new
           @body = options[:body] || StringIO.new
-          @raw_stream = options[:raw_stream] || StringIO.new
           @listeners = Hash.new { |h,k| h[k] = [] }
           @complete = false
           @done = nil
@@ -26,12 +25,6 @@ module Seahorse
 
         # @return [StandardError, nil]
         attr_reader :error
-
-        # Raw binary stream data used for eventstream only
-        # body is concatenated parsed event structs in StringIO
-        #
-        # @return [IO]
-        attr_accessor :raw_stream
 
         # @return [IO]
         def body
@@ -66,9 +59,6 @@ module Seahorse
         # @param [string] chunk
         def signal_data(chunk)
           unless chunk == ''
-            # record raw binary stream for eventstream
-            # in case error happens, body is used for emit/track events
-            @raw_stream.write(chunk)
             @body.write(chunk)
             emit(:data, chunk)
           end